requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
dateparser==1.2.0
fonttools==4.49.0
frozenlist==1.4.1
idna==3.6
importlib-resources==6.1.1
kiwisolver==1.4.5
matplotlib==3.8.3
multidict==6.0.5
numpy==1.26.4
packaging==23.2
pandas==2.2.1
pillow==10.2.0
pycryptodome==3.20.0
pyparsing==3.1.1
pyTelegramBotAPI==4.16.1
python-binance==1.0.19
python-dateutil==2.8.2
pytz==2024.1
regex==2023.12.25
requests==2.31.0
seaborn==0.13.2
six==1.16.0
telebot==0.0.5
tzdata==2024.1
tzlocal==5.2
ujson==5.9.0
urllib3==2.2.1
websockets==12.0
yarl==1.9.4
zipp==3.17.0
